Remove control plane
