+++ Microsoft.Health.Fhir.sln
@@ -42,11 +42,7 @@ Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Solution Items", "Solution 
 		testauthenvironment.json = testauthenvironment.json
 	EndProjectSection
 EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Microsoft.Health.ControlPlane.Core", "src\Microsoft.Health.ControlPlane.Core\Microsoft.Health.ControlPlane.Core.csproj", "{E2403F57-E9DE-419C-A64D-C45B3C683762}"
-EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Microsoft.Health.CosmosDb", "src\Microsoft.Health.CosmosDb\Microsoft.Health.CosmosDb.csproj", "{AFE75E20-714A-444C-928D-490CB8469B58}"
-EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Microsoft.Health.ControlPlane.CosmosDb", "src\Microsoft.Health.ControlPlane.CosmosDb\Microsoft.Health.ControlPlane.CosmosDb.csproj", "{11AB85A3-575C-40E7-A36C-66504B57F07A}"
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Microsoft.Health.Abstractions", "src\Microsoft.Health.Abstractions\Microsoft.Health.Abstractions.csproj", "{E0CD7A0A-116D-4612-83BE-0DE036DADDF1}"
 EndProject
@@ -58,11 +54,9 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "SmartLauncher", "samples\ap
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Microsoft.Health.CosmosDb.UnitTests", "src\Microsoft.Health.CosmosDb.UnitTests\Microsoft.Health.CosmosDb.UnitTests.csproj", "{A6EDEC42-6828-4ADE-A0B4-2042B4AC65F7}"
 EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Microsoft.Health.ControlPlane.Core.UnitTests", "src\Microsoft.Health.ControlPlane.Core.UnitTests\Microsoft.Health.ControlPlane.Core.UnitTests.csproj", "{A906D1FB-9282-46A0-ADDC-46726B2C1378}"
-EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Microsoft.Health.Fhir.SqlServer.Api", "src\Microsoft.Health.Fhir.SqlServer.Api\Microsoft.Health.Fhir.SqlServer.Api.csproj", "{1C93EE4A-2704-4E22-8A17-1CCC73F3807E}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Microsoft.Health.Fhir.SqlServer.Api.UnitTests", "src\Microsoft.Health.Fhir.SqlServer.Api.UnitTests\Microsoft.Health.Fhir.SqlServer.Api.UnitTests.csproj", "{6746C0B9-3B51-4118-8C5D-9592359F7A71}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Microsoft.Health.Fhir.SqlServer.Api.UnitTests", "src\Microsoft.Health.Fhir.SqlServer.Api.UnitTests\Microsoft.Health.Fhir.SqlServer.Api.UnitTests.csproj", "{6746C0B9-3B51-4118-8C5D-9592359F7A71}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -126,18 +120,10 @@ Global
 		{CA276939-8071-4734-9FE4-ADC825B72116}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{CA276939-8071-4734-9FE4-ADC825B72116}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{CA276939-8071-4734-9FE4-ADC825B72116}.Release|Any CPU.Build.0 = Release|Any CPU
-		{E2403F57-E9DE-419C-A64D-C45B3C683762}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{E2403F57-E9DE-419C-A64D-C45B3C683762}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{E2403F57-E9DE-419C-A64D-C45B3C683762}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{E2403F57-E9DE-419C-A64D-C45B3C683762}.Release|Any CPU.Build.0 = Release|Any CPU
 		{AFE75E20-714A-444C-928D-490CB8469B58}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{AFE75E20-714A-444C-928D-490CB8469B58}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{AFE75E20-714A-444C-928D-490CB8469B58}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{AFE75E20-714A-444C-928D-490CB8469B58}.Release|Any CPU.Build.0 = Release|Any CPU
-		{11AB85A3-575C-40E7-A36C-66504B57F07A}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{11AB85A3-575C-40E7-A36C-66504B57F07A}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{11AB85A3-575C-40E7-A36C-66504B57F07A}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{11AB85A3-575C-40E7-A36C-66504B57F07A}.Release|Any CPU.Build.0 = Release|Any CPU
 		{E0CD7A0A-116D-4612-83BE-0DE036DADDF1}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{E0CD7A0A-116D-4612-83BE-0DE036DADDF1}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{E0CD7A0A-116D-4612-83BE-0DE036DADDF1}.Release|Any CPU.ActiveCfg = Release|Any CPU
@@ -150,10 +136,6 @@ Global
 		{A6EDEC42-6828-4ADE-A0B4-2042B4AC65F7}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{A6EDEC42-6828-4ADE-A0B4-2042B4AC65F7}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{A6EDEC42-6828-4ADE-A0B4-2042B4AC65F7}.Release|Any CPU.Build.0 = Release|Any CPU
-		{A906D1FB-9282-46A0-ADDC-46726B2C1378}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{A906D1FB-9282-46A0-ADDC-46726B2C1378}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{A906D1FB-9282-46A0-ADDC-46726B2C1378}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{A906D1FB-9282-46A0-ADDC-46726B2C1378}.Release|Any CPU.Build.0 = Release|Any CPU
 		{1C93EE4A-2704-4E22-8A17-1CCC73F3807E}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{1C93EE4A-2704-4E22-8A17-1CCC73F3807E}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{1C93EE4A-2704-4E22-8A17-1CCC73F3807E}.Release|Any CPU.ActiveCfg = Release|Any CPU
@@ -181,19 +163,16 @@ Global
 		{E2C2B117-32CD-449F-A4A8-4C8BA874D693} = {8AD2A324-DAB5-4380-94A5-31F7D817C384}
 		{614DE061-52A0-4471-A3F1-50C9C907FFDB} = {8AD2A324-DAB5-4380-94A5-31F7D817C384}
 		{CA276939-8071-4734-9FE4-ADC825B72116} = {B70945F4-01A6-4351-955B-C4A2943B5E3B}
-		{E2403F57-E9DE-419C-A64D-C45B3C683762} = {8AD2A324-DAB5-4380-94A5-31F7D817C384}
 		{AFE75E20-714A-444C-928D-490CB8469B58} = {8AD2A324-DAB5-4380-94A5-31F7D817C384}
-		{11AB85A3-575C-40E7-A36C-66504B57F07A} = {8AD2A324-DAB5-4380-94A5-31F7D817C384}
 		{E0CD7A0A-116D-4612-83BE-0DE036DADDF1} = {8AD2A324-DAB5-4380-94A5-31F7D817C384}
 		{80E8C5D3-E02C-49C0-969B-0C3A02FA1FF4} = {B5F2D2DF-D0C7-4861-8259-F6A041DB9854}
 		{49CA3B9D-7A22-43E0-A434-B0EC38D119A5} = {80E8C5D3-E02C-49C0-969B-0C3A02FA1FF4}
 		{A6EDEC42-6828-4ADE-A0B4-2042B4AC65F7} = {8AD2A324-DAB5-4380-94A5-31F7D817C384}
-		{A906D1FB-9282-46A0-ADDC-46726B2C1378} = {8AD2A324-DAB5-4380-94A5-31F7D817C384}
 		{1C93EE4A-2704-4E22-8A17-1CCC73F3807E} = {8AD2A324-DAB5-4380-94A5-31F7D817C384}
 		{6746C0B9-3B51-4118-8C5D-9592359F7A71} = {8AD2A324-DAB5-4380-94A5-31F7D817C384}
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
-		RESX_SortFileContentOnSave = True
 		SolutionGuid = {E370FB31-CF95-47D1-B1E1-863A77973FF8}
+		RESX_SortFileContentOnSave = True
 	EndGlobalSection
 EndGlobal
