--- src/Microsoft.Health.ControlPlane.Core/Features/Persistence/UpsertOutcomeType.cs
@@ -1,13 +1,0 @@
-﻿// -------------------------------------------------------------------------------------------------
-// Copyright (c) Microsoft Corporation. All rights reserved.
-// Licensed under the MIT License (MIT). See LICENSE in the repo root for license information.
-// -------------------------------------------------------------------------------------------------
-
-namespace Microsoft.Health.ControlPlane.Core.Features.Persistence
-{
-    public enum UpsertOutcome
-    {
-        Created,
-        Updated,
-    }
-}
