--- src/Microsoft.Health.ControlPlane.Core/Features/Rbac/ResourcePermission.cs
@@ -1,24 +1,0 @@
-﻿// -------------------------------------------------------------------------------------------------
-// Copyright (c) Microsoft Corporation. All rights reserved.
-// Licensed under the MIT License (MIT). See LICENSE in the repo root for license information.
-// -------------------------------------------------------------------------------------------------
-
-using System.Collections.Generic;
-
-namespace Microsoft.Health.ControlPlane.Core.Features.Rbac
-{
-    public class ResourcePermission
-    {
-        public ResourcePermission()
-        {
-            Actions = new List<ResourceAction>();
-        }
-
-        public ResourcePermission(IList<ResourceAction> resourceActions)
-        {
-            Actions = resourceActions;
-        }
-
-        public IList<ResourceAction> Actions { get; }
-    }
-}
