--- src/Microsoft.Health.ControlPlane.CosmosDb/AssemblyInfo.cs
@@ -1,10 +1,0 @@
-﻿// -------------------------------------------------------------------------------------------------
-// Copyright (c) Microsoft Corporation. All rights reserved.
-// Licensed under the MIT License (MIT). See LICENSE in the repo root for license information.
-// -------------------------------------------------------------------------------------------------
-
-using System.Resources;
-using System.Runtime.CompilerServices;
-
-[assembly: InternalsVisibleTo("Microsoft.Health.ControlPlane.CosmosDb.UnitTests")]
-[assembly: NeutralResourcesLanguage("en-us")]
