--- src/Microsoft.Health.ControlPlane.CosmosDb/Constants.cs
@@ -1,12 +1,0 @@
-﻿// -------------------------------------------------------------------------------------------------
-// Copyright (c) Microsoft Corporation. All rights reserved.
-// Licensed under the MIT License (MIT). See LICENSE in the repo root for license information.
-// -------------------------------------------------------------------------------------------------
-
-namespace Microsoft.Health.ControlPlane.CosmosDb
-{
-    internal static class Constants
-    {
-        public const string CollectionConfigurationName = "controlPlaneCosmosDb";
-    }
-}
