--- src/Microsoft.Health.ControlPlane.CosmosDb/Features/Storage/Versioning/IControlPlaneCollectionUpdater.cs
@@ -1,13 +1,0 @@
-﻿// -------------------------------------------------------------------------------------------------
-// Copyright (c) Microsoft Corporation. All rights reserved.
-// Licensed under the MIT License (MIT). See LICENSE in the repo root for license information.
-// -------------------------------------------------------------------------------------------------
-
-using Microsoft.Health.CosmosDb.Features.Storage.Versioning;
-
-namespace Microsoft.Health.ControlPlane.CosmosDb.Features.Storage.Versioning
-{
-    public interface IControlPlaneCollectionUpdater : ICollectionUpdater
-    {
-    }
-}
