--- src/Microsoft.Health.Fhir.Api/Modules/ControlPlaneModule.cs
@@ -1,29 +1,0 @@
-﻿// -------------------------------------------------------------------------------------------------
-// Copyright (c) Microsoft Corporation. All rights reserved.
-// Licensed under the MIT License (MIT). See LICENSE in the repo root for license information.
-// -------------------------------------------------------------------------------------------------
-
-using EnsureThat;
-using Microsoft.Extensions.DependencyInjection;
-using Microsoft.Health.ControlPlane.Core.Features.Rbac;
-using Microsoft.Health.Extensions.DependencyInjection;
-using Microsoft.Health.Fhir.Api.Configs;
-
-namespace Microsoft.Health.Fhir.Api.Modules
-{
-    public class ControlPlaneModule : IStartupModule
-    {
-        public ControlPlaneModule(FhirServerConfiguration fhirServerConfiguration)
-        {
-            EnsureArg.IsNotNull(fhirServerConfiguration, nameof(fhirServerConfiguration));
-        }
-
-        /// <inheritdoc />
-        public void Load(IServiceCollection services)
-        {
-            EnsureArg.IsNotNull(services, nameof(services));
-
-            services.AddScoped<IRbacService, RbacService>();
-        }
-    }
-}
