+++ src/Microsoft.Health.Fhir.Web/appsettings.json
@@ -38,12 +38,6 @@
       }
     }
   },
-  "ControlPlane": {
-    "CosmosDb": {
-      "CollectionId": "controlPlane",
-      "InitialCollectionThroughPut": null
-    }
-  },
   "CosmosDb": {
     "Host": null,
     "Key": null,
